+++ README.md
@@ -31,8 +31,10 @@ In addition, take note of the API Key value for your account. This value
 must be passed as an INORBIT_ICM_KEY environment variable to the
 service.
 
-Add INORBIT_BASE_URL environment variable with InOrbit base url. If you don't set up this variable
-it will use https://control.inorbit.ai as default.
+## Advanced Configuration
+
+If you are using a specific InOrbit deployment then add INORBIT_BASE_URL environment variable with your InOrbit base url.
+If you don't set up this variable it will use https://control.inorbit.ai as default.
 
 ## Running
